Ship Tools code rewrite and fix.
* Finally took the time to fix this, welder/grinders/drills
should work as expected now,

* BuildAndRapair has it's own logic so it's will be ignored.

* Welder will not count it's self as target, as weld projections as should.
* Grinder will collect inventory as should.

All 3 blocks have custom configurations via gui fro speed and drill tick rate.

Thanks to SlimeRadio for help.

+++ DePatch/ShipTools/DrillSettings.cs
@@ -16,7 +16,7 @@ namespace DePatch.ShipTools
 
         public bool DisableRightClick { get; set; }
 
-        public float TickRate { get; set; } = 90f;
+        public int TickRate { get; set; } = 90;
 
         public static string Serialize(DrillSettings settings) => string.Format("{0}:{1}:{2}:{3}",
                                                                   settings.Subtype,
@@ -33,7 +33,7 @@ namespace DePatch.ShipTools
                 Subtype = array[0],
                 DisableRightClick = bool.Parse(array[1]),
                 Mode = (DrillingMode)Enum.Parse(typeof(DrillingMode), array[2]),
-                TickRate = float.Parse(array[3])
+                TickRate = int.Parse(array[3])
             };
         }
